safari-534.53-branch/LayoutTests/http/tests/security/xssAuditor/no-protection-script-tag-expected.txt

ALERT: /PASS/
This script should run because XSS protection is disabled.
